Feature: authorization handling

+++ UserAccess.Application/Users/RemoveUser/RemoveUserCommandHandler.cs
@@ -9,10 +9,12 @@ namespace UserAccess.Application.Users.RemoveUser;
 internal sealed class RemoveUserCommandHandler : ICommandRequestHandler<RemoveUserCommand, ErrorOr<Unit>>
 {
     private readonly IUserRepository _userRepository;
+    private readonly AuthorizationService _authorizationService;
 
-    public RemoveUserCommandHandler(IUserRepository userRepository)
+    public RemoveUserCommandHandler(IUserRepository userRepository, AuthorizationService authorizationService)
     {
         _userRepository = userRepository;
+        _authorizationService = authorizationService;
     }
 
     public async Task<ErrorOr<Unit>> Handle(RemoveUserCommand request, CancellationToken cancellationToken)
@@ -22,6 +24,13 @@ internal sealed class RemoveUserCommandHandler : ICommandRequestHandler<RemoveUs
         if (user is null)
         {
             return UserErrorsCodes.NotFound;
+        }
+
+        var authorizationService = _authorizationService.IsUserAuthorized(user.Id.Value);
+
+        if (authorizationService.IsError && _authorizationService.IsAdmin() is false)
+        {
+            return UserErrorsCodes.CannotRemove;
         }
 
         await _userRepository.RemoveAsync(user);
